move llmconfig beneath internal/cmd/dagger/
