+++ internal/cmd/dagger/llmconfig/suggestions_test.go
